Work on Magento native 2fa compatibility
