+++ etc/adminhtml/events.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:noNamespaceSchemaLocation="urn:magento:framework:Event/etc/events.xsd">
+    <event name="controller_action_predispatch">
+        <observer name="admin_restriction_check"
+                  instance="MSP\AdminRestriction\Observer\CheckAdminRestriction"/>
+    </event>
+</config>
